requirements.txt
# Runtime dependencies
Flask==1.1.1
Flask-API==1.1
Flask_API==1.1
Flask-SQLAlchemy==2.4.1
psycopg2-binary==2.8.3
python-dotenv==0.10.3

# Testing
nose==1.3.7
rednose==1.3.0
pinocchio==0.4.2
factory-boy==2.12.0
coverage==4.5.4
pylint>=2.4.1